fix(server): prevent race conditions when creating db

+++ scripts/client/adapter.js
@@ -105,7 +105,7 @@ Adapter.prototype._systemDB = function () {
   return this._sysDB;
 };
 
-Adapter.prototype._resolveAfterDatabaseCreated = function (dbName, originatingDoc) {
+Adapter.prototype._resolveAfterDatabaseCreated = function (dbName, originatingDoc, ts) {
   return new Promise(function (resolve) {
     // When creating a DB, the delta is id-less and so that cannot use an id to reconcile the
     // local doc. Instead we listen for a new doc on the parent collection and then delete the
@@ -114,21 +114,64 @@ Adapter.prototype._resolveAfterDatabaseCreated = function (dbName, originatingDo
     // corresponds to the creating delta id.
     originatingDoc._col.on('doc:create', function (doc) {
       var data = doc.get();
-      if (data[clientUtils.DB_ATTR_NAME] && data[clientUtils.DB_ATTR_NAME] === dbName) {
+      // There could have been DBs with the same name created before so we need to check the
+      // timestamp
+// if (data[clientUtils.DB_ATTR_NAME] && data[clientUtils.DB_ATTR_NAME] === dbName) {
+// console.log('%%%%%%%%%%%%%%%%%%%%%%%      ', doc._dat.recordedAt.getTime(), ts.getTime());
+
+      if (data[clientUtils.DB_ATTR_NAME] && data[clientUtils.DB_ATTR_NAME] === dbName
+        && doc._dat.recordedAt.getTime() >= ts.getTime()) {
+// console.log('Adapter.prototype._resolveAfterDatabaseCreated, ts=', ts, 'doc=', doc);
+// process.exit(1);
         resolve(originatingDoc._destroyLocally());
       }
     });
   });
 };
 
+// Adapter.prototype._resolveAfterDatabaseCreated = function (dbName, originatingDoc, ts) {
+//   return new Promise(function (resolve) {
+//     // When creating a DB, the delta is id-less and so that cannot use an id to reconcile the
+//     // local doc. Instead we listen for a new doc on the parent collection and then delete the
+//     // local doc that was used to originate the delta so that we don't attempt to create the DB
+//     // again. TODO: Another option for the future could be to create an id in the doc that
+//     // corresponds to the creating delta id.
+//     originatingDoc._col.on('attr:record', function (attr, doc) {
+// console.log('Adapter.prototype._resolveAfterDatabaseCreated, attr=', attr, 'doc=', doc);
+// process.exit(1);
+// //       var data = doc.get();
+// //       if (data[clientUtils.DB_ATTR_NAME] && data[clientUtils.DB_ATTR_NAME] === dbName) {
+// // console.log('Adapter.prototype._resolveAfterDatabaseCreated, ts=', ts, 'doc=', doc);
+// //         resolve(originatingDoc._destroyLocally());
+// //       }
+//     });
+//   });
+// };
+
 Adapter.prototype._createDatabase = function (dbName) {
-  var self = this;
+  var self = this, ts = new Date();
   return self._systemDB()._createDatabase(dbName).then(function (doc) {
-    return self._resolveAfterDatabaseCreated(dbName, doc);
+    return self._resolveAfterDatabaseCreated(dbName, doc, ts);
   });
 };
 
-Adapter.prototype._resolveAfterDatabaseDestroyed = function (dbName, originatingDoc) {
+// Adapter.prototype._resolveAfterDatabaseDestroyed = function (dbName, originatingDoc) {
+//   return new Promise(function (resolve) {
+//     // When creating a DB, the delta is id-less and so we cannot use an id to reconcile the
+//     // local doc. Instead we listen for a doc:destroy event on the parent collection and then delete the
+//     // local doc that was used to originate the delta so that we don't attempt to destroy the DB
+//     // again. TODO: Another option for the future could be to create an id in the doc that
+//     // corresponds to the destroying delta id.
+//     originatingDoc._col.on('doc:destroy', function (doc) {
+//       var data = doc.get();
+//       if (data[clientUtils.DB_ATTR_NAME] && data[clientUtils.DB_ATTR_NAME] === dbName) {
+//         resolve(originatingDoc._destroyLocally());
+//       }
+//     });
+//   });
+// };
+
+Adapter.prototype._resolveAfterDatabaseDestroyed = function (dbName, originatingDoc, ts) {
   return new Promise(function (resolve) {
     // When creating a DB, the delta is id-less and so we cannot use an id to reconcile the
     // local doc. Instead we listen for a doc:destroy event on the parent collection and then delete the
@@ -137,7 +180,12 @@ Adapter.prototype._resolveAfterDatabaseDestroyed = function (dbName, originating
     // corresponds to the destroying delta id.
     originatingDoc._col.on('doc:destroy', function (doc) {
       var data = doc.get();
-      if (data[clientUtils.DB_ATTR_NAME] && data[clientUtils.DB_ATTR_NAME] === dbName) {
+// if (data[clientUtils.DB_ATTR_NAME] && data[clientUtils.DB_ATTR_NAME] === dbName) {
+// console.log('Adapter.prototype._resolveAfterDatabaseDestroyed, ts=', ts, 'doc=', doc);
+     if (data[clientUtils.DB_ATTR_NAME] && data[clientUtils.DB_ATTR_NAME] === dbName
+       && doc._dat.destroyedAt.getTime() >= ts.getTime()) {
+        // There could have been DBs with the same name destroyed before so we need to check the
+        // timestamp
         resolve(originatingDoc._destroyLocally());
       }
     });
@@ -145,7 +193,7 @@ Adapter.prototype._resolveAfterDatabaseDestroyed = function (dbName, originating
 };
 
 Adapter.prototype._destroyDatabase = function (dbName) {
-  var self = this;
+  var self = this, ts = new Date();
 
   // If the db exists then close it first!! I don't think we have to worry about a race condition
   // where the client re-creates this DB while trying to destroy as the destroy will just fail as
@@ -169,7 +217,7 @@ Adapter.prototype._destroyDatabase = function (dbName) {
   return promise.then(function () {
     return self._systemDB()._destroyDatabase(dbName);
   }).then(function (doc) {
-    return self._resolveAfterDatabaseDestroyed(dbName, doc);
+    return self._resolveAfterDatabaseDestroyed(dbName, doc, ts);
   });
 };
